[Backend] Add most active user endpoint Fixes #510
