+++ web_app/api/serializers/leaderboard.py
@@ -1,10 +1,11 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, Depends
 from pydantic import BaseModel
 from typing import List
-from web_app.db.crud import PositionDBConnector
+from sqlalchemy.orm import Session
+from web_app.db.crud.leaderboard import LeaderboardCRUD
+from web_app.db.session import get_db
 
 router = APIRouter()
-position_db_connector = PositionDBConnector()
 
 class UserLeaderboardItem(BaseModel):
     wallet_id: str
@@ -20,9 +21,10 @@ class UserLeaderboardResponse(BaseModel):
     summary="Get user leaderboard",
     response_description="Returns the top 10 users ordered by closed/opened positions.",
 )
-async def get_user_leaderboard() -> UserLeaderboardResponse:
+async def get_user_leaderboard(db: Session = Depends(get_db)) -> UserLeaderboardResponse:
     """
     Get the top 10 users ordered by closed/opened positions.
     """
-    leaderboard_data = position_db_connector.get_top_users_by_positions()
+    leaderboard_crud = LeaderboardCRUD(db)
+    leaderboard_data = leaderboard_crud.get_top_users_by_positions()
     return UserLeaderboardResponse(leaderboard=leaderboard_data)
