+++ web_app/db/crud/position.py
@@ -103,35 +103,7 @@ class PositionDBConnector(UserDBConnector):
             except SQLAlchemyError as e:
                 logger.error(f"Failed to retrieve positions: {str(e)}")
                 return []
-            
-    def get_top_users_by_positions(self) -> list[dict]:
-        """
-        Retrieves the top 10 users ordered by closed/opened positions.
-        :return: List of dictionaries containing wallet_id and positions_number.
-        """
-        with self.Session() as db:
-            try:
-                results = (
-                    db.query(
-                        User.wallet_id,
-                        func.count(Position.id).label("positions_number")
-                    )
-                    .join(Position, Position.user_id == User.id)
-                    .group_by(User.wallet_id)
-                    .order_by(func.count(Position.id).desc())
-                    .limit(10)
-                    .all()
-                )
-
-                return [
-                    {"wallet_id": result.wallet_id, "positions_number": result.positions_number}
-                    for result in results
-                ]
-
-            except SQLAlchemyError as e:
-                logger.error(f"Error retrieving top users by positions: {e}")
-                return []
-            
+      
     def get_all_positions_by_wallet_id(
         self, wallet_id: str, start: int, limit: int
     ) -> list:
@@ -527,3 +499,5 @@ class PositionDBConnector(UserDBConnector):
                 .all()
             )
             return extra_deposits
+
+
